# Optimize IsaacLab when no cameras