This PR optimize the launching speed and VRAM consumption of isaaclab handler when no camera.

This PR is experimental and is continually updating.


### Results
- The launch time reduced from 6.5s to 3.75s
- The VRAM usage reduced from ~4GB to ~3GB
- The per step time remain the same: ~0.12s

Tested on RTX4060 + i7-14700F

### How to reproduce
Remove the default camera in `replay_demo.py` from 
```python
scenario = ScenarioCfg(
    ...
    cameras=[camera],
    ...
)
```
to
```python
scenario = ScenarioCfg(
    ...
    cameras=[],
    ...
)
```

And test using
```bash
LOGURU_LEVEL=TRACE python metasim/scripts/replay_demo.py --task=stack_cube --sim=isaaclab --num-envs=960 --headless
```

### Discussion
It is worth-noting that the test using IsaacGym handler with same setup:
```
python metasim/scripts/replay_demo.py --task=stack_cube --sim=isaacgym --num-envs=960 --headless
```
achieves ~0.022s per step, which is 500% faster than IsaacLab. Still figuring out why IsaacLab is so slow.

### Other attempt to speed up IsaacLab
- Tried using `torch.inference_mode` but not faster