@Fisher-Wang 2025-08-27
It turns out this optimization is insignificant compared to the flexibility.